+++ core/types/Interfaces.ts
@@ -182,7 +182,7 @@ export interface AppOptionSpec {
     reloadRequired?: boolean;
 
     /** Optional flag to omit displaying option. */
-    omit?: boolean;
+    omit?: boolean|(() => boolean);
 }
 
 /**
@@ -219,7 +219,7 @@ export interface TrackOptions {
     elapsed?: number;
 
     /** Optional flag to omit sending message. */
-    omit?: boolean;
+    omit?: boolean|(() => boolean);
 }
 
 
@@ -255,7 +255,7 @@ export interface MenuItem {
     hidden?: boolean;
 
     /** True to skip this item. May be set dynamically via prepareFn. Alias for hidden.  */
-    omit?: boolean;
+    omit?: boolean|(() => boolean);
 }
 
 /**